Add files via upload
TASK-3 CAR PRICE PREDICTION USING MACHINE LEARNING 
DOMAIN : DATA SCIENCE

The price of a car depends on a lot of factors like the good will of the brand of the car, features of the car, horoscope and the milage it gives and many more. Car price prediction is one of the major research areas in machine learning. So if you want to learn how to train a car price prediction model then this project is for you.