Add files via upload
Task-4: EMAIL SPAM DETECTION WITH MACHINE LEARNING
DOMAIN : DATA SCIENCE

The Machine Learning Model use the train data to train itself, so as to classify the messages/emails into 2 categories namely: Spam and Ham and then provide the spam prediction results.
The Classifier used is : MultinomialNB (Naïve Bayes) Classifier,
it assumes that the features represent discrete counts, such as word frequencies in a text classification task. This variant is suitable for data with a limited number of possible values.
The accuracy is measured by plotting the Confusion Matrix.